server/requirements.txt
fastapi==0.111.1
uvicorn==0.30.1
pillow==10.4.0
numpy<2.0
torch==2.1.2
torchvision==0.16.2
diffusers==0.30.0
transformers==4.36.2
accelerate==0.25.0
opencv-python-headless==4.10.0.84
huggingface-hub==0.19.4
